Move request context and modify initialization
